--- .github/workflows/notify-amount-transferred--test.yaml
@@ -1,61 +1,0 @@
-name: notify-amount-transferred / test
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - .github/workflows/notify-amount-transferred--test.yaml
-      - notify-amount-transferred/**
-  pull_request:
-    branches: [main]
-    paths:
-      - .github/workflows/notify-amount-transferred--test.yaml
-      - notify-amount-transferred/**
-
-defaults:
-  run:
-    working-directory: notify-amount-transferred
-
-jobs:
-  lint:
-    needs: fix
-    runs-on: ubuntu-latest
-    env:
-      GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-    steps:
-      - uses: actions/checkout@v3
-      - uses: aquaproj/aqua-installer@v1.2.0
-        with:
-          aqua_version: v1.38.0
-      - uses: actions/setup-node@v3
-        with:
-          node-version: "16"
-      - run: github-comment exec -- yarn install --frozen-lockfile
-      - run: github-comment exec -- yarn lint
-
-  fix:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: aquaproj/aqua-installer@v1.2.0
-        with:
-          aqua_version: v1.38.0
-      - run: github-comment exec --token ${{ secrets.GITHUB_TOKEN }} -- yarn install --frozen-lockfile
-      - run: github-comment exec --token ${{ secrets.GITHUB_TOKEN }} -- yarn fix
-      - uses: int128/update-generated-files-action@v2
-
-  test:
-    runs-on: ubuntu-latest
-    env:
-      GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-    steps:
-      - uses: actions/checkout@v3
-      - uses: aquaproj/aqua-installer@v1.2.0
-        with:
-          aqua_version: v1.38.0
-      - uses: actions/setup-node@v3
-        with:
-          node-version: "16"
-      - run: github-comment exec -- yarn install --frozen-lockfile
-      - run: github-comment exec -- yarn test
