--- notify-amount-transferred/.clasp.json
@@ -1,4 +1,0 @@
-{
-  "scriptId": "1ynV4jq399FRnTi4BNqDBRyy6x9poROgt7U7_pIzizEpnXqmw5zqqpgh2",
-  "rootDir": "src"
-}
