--- notify-amount-transferred/.prettierrc
@@ -1,8 +1,0 @@
-{
-    "printWidth": 100,
-    "tabWidth": 2,
-    "singleQuote": true,
-    "semi": true,
-    "quoteProps": "preserve",
-    "bracketSpacing": false
-}
