--- notify-amount-transferred/package.json
@@ -1,26 +1,0 @@
-{
-  "dependencies": {
-    "@types/google-apps-script": "1.0.62"
-  },
-  "devDependencies": {
-    "@types/jest": "27.5.2",
-    "@typescript-eslint/eslint-plugin": "5.59.2",
-    "@typescript-eslint/parser": "5.59.2",
-    "eslint": "8.39.0",
-    "eslint-config-prettier": "8.8.0",
-    "jest": "28.1.3",
-    "npm-run-all": "4.1.5",
-    "prettier": "2.8.8",
-    "ts-jest": "28.0.8",
-    "typescript": "4.9.5"
-  },
-  "scripts": {
-    "test": "jest",
-    "fix": "run-s fix:prettier fix:eslint",
-    "fix:prettier": "prettier --write src",
-    "fix:eslint": "eslint src --ext .ts --fix",
-    "lint": "run-p lint:prettier lint:eslint",
-    "lint:prettier": "prettier --check src",
-    "lint:eslint": "eslint src --ext .ts"
-  }
-}
