+++ notify-common-expense/.editorconfig
@@ -1,0 +1,7 @@
+root = true
+[*]
+indent_style = space
+indent_size = 2
+end_of_line = lf
+charset = utf-8
+insert_final_newline = true
