+++ notify-common-expense/.eslintignore
@@ -1,0 +1,4 @@
+**/node_modules/*
+build/*
+**/dist/*
+testing
