+++ notify-common-expense/.gitignore
@@ -1,0 +1,7 @@
+node_modules
+dist/
+build/
+.DS_Store
+.clasp*.json
+*.tgz
+.vscode/
