+++ notify-common-expense/.prettierrc.json
@@ -1,0 +1,7 @@
+{
+  "singleQuote": true,
+  "trailingComma": "es5",
+  "arrowParens": "avoid",
+  "quoteProps": "consistent",
+  "tabWidth": 2
+}
