+++ notify-common-expense/jest.config.json
@@ -1,0 +1,9 @@
+{
+  "preset": "ts-jest",
+  "testEnvironment": "node",
+  "modulePathIgnorePatterns": ["<rootDir>/dist/", "<rootDir>/build/"],
+  "testPathIgnorePatterns": ["<rootDir>/dist/", "<rootDir>/build/"],
+  "moduleNameMapper": {
+    "(.+)\\.js": "$1"
+  }
+}
