+++ notify-common-expense/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "notify-common-expense",
+  "version": "0.0.0",
+  "description": "",
+  "main": "build/index.js",
+  "license": "Apache-2.0",
+  "keywords": [],
+  "scripts": {
+    "clean": "rimraf build dist",
+    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
+    "lint": "npm run license && eslint --fix --no-error-on-unmatched-pattern src/ test/",
+    "bundle": "rollup --no-treeshake -c rollup.config.mjs",
+    "build": "npm run clean && npm run bundle && ncp appsscript.json dist/appsscript.json",
+    "license": "license-check-and-add add -f license-config.json",
+    "test": "jest test/ --passWithNoTests --detectOpenHandles",
+    "deploy": "npm run lint && npm run test && npm run build && ncp .clasp-dev.json .clasp.json && clasp push -f",
+    "deploy:prod": "npm run lint && npm run test && npm run build && ncp .clasp-prod.json .clasp.json && clasp push"
+  },
+  "engines": {
+    "node": ">=22"
+  },
+  "dependencies": {
+    "@google/clasp": "^3.0.6-alpha",
+    "@types/google-apps-script": "^2.0.3",
+    "@types/jest": "^30.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "eslint": "^8.57.1",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.4",
+    "gts": "^6.0.2",
+    "jest": "^30.1.3",
+    "license-check-and-add": "^4.0.5",
+    "ncp": "^2.0.0",
+    "prettier": "^3.6.2",
+    "rimraf": "^6.0.1",
+    "rollup": "^4.52.0",
+    "rollup-plugin-cleanup": "^3.2.1",
+    "rollup-plugin-license": "^3.6.0",
+    "rollup-plugin-prettier": "^4.1.2",
+    "rollup-plugin-typescript2": "^0.36.0",
+    "ts-jest": "^29.4.4",
+    "typescript": "^5.9.2"
+  }
+}
