+++ notify-common-expense/src/property.ts
@@ -1,0 +1,32 @@
+/**
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+export interface PropertyStore {
+  get(key: string): string | undefined;
+}
+
+export class ScriptPropertyStore implements PropertyStore {
+  constructor(
+    private props: GoogleAppsScript.Properties.Properties = PropertiesService.getScriptProperties()
+  ) {}
+
+  get(key: string): string | undefined {
+    const val = this.props.getProperty(key);
+    if (val == null) {
+      return undefined;
+    }
+    return val;
+  }
+}
